충남대 BE_김민지 6주차 과제(1단계,2단계) (#217)
* feat : step0 진행

* refactor : password 검증 로직 엔티티에 저장

* test : test코드 안돌아가는 것 수정

* test : menuname 테스트

* refactor : Menu이름 VO 생성

* refactor : memberController api변경

* refactor : 카카오 로그인 시 프로젝트 내의 토큰 반환

* refactor : 카테고리 api 변경

* refactor : api 변경

* refactor : 로그인기능 명세대로 수정

* refactor : Category 기능 명세대로 수정

* refactor : 상품 api 명세대로 수정

* refactor : 로그인 api 수정

* refactor : 메뉴 api 수정

* refactor : 옵션 api 수정

* refactor : 옵션 api 수정

* refactor : order api 수정

* refactor : menu api 수정

* refactor : wishlist api 수정

* test : test코드 수정

* refactor : 옵션명, 수량 유효성 체크 엔티티 생성

* test : 옵션명, 수량 유효성 체크 테스트 생성

* refactor : 불필요한 import 제거

* refactor : 패키지명 변경

* Update application.yml

* refactor : swagger 제목 변경

* refactor : api명세에 맞게 수정

* refactor : cors 허용

+++ src/main/java/gift/service/WishListService.java
@@ -1,7 +1,7 @@
 package gift.service;
 
-import gift.domain.WishList;
-import gift.domain.WishListResponse;
+import gift.domain.WishListDomain.WishList;
+import gift.domain.WishListDomain.WishListResponse;
 import gift.repository.WishListRepository;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -17,23 +17,27 @@ public class WishListService {
         this.wishListRepository = wishListRepository;
     }
 
-    public void save(WishList wishList) {
-        wishListRepository.save(wishList);
+    public WishListResponse save(WishList wishList) {
+        return mapWishListToWishListResponse(wishListRepository.save(wishList));
     }
 
     public List<WishListResponse> findById(String jwtId, Pageable pageable) {
         List<WishList> wishLists = wishListRepository.findByMemberId(jwtId, pageable);
         return wishLists.stream()
-                .map(this::MapWishListToWishListResponse)
+                .map(this::mapWishListToWishListResponse)
                 .collect(Collectors.toList());
     }
 
-    public void delete(Long Id) {
-        wishListRepository.deleteById(Id);
+    public void delete(String jwtId, Long Id) throws IllegalAccessException {
+        List<WishList> wishLists = wishListRepository.findByMemberId(jwtId);
+        if(wishLists.contains(wishListRepository.findById(Id))) {
+            wishListRepository.deleteById(Id);
+        }
+        throw new IllegalAccessException("해당 위시에 대한 권한이 없습니다.");
     }
 
-    public WishListResponse MapWishListToWishListResponse(WishList wishList) {
-        return new WishListResponse(wishList.getId(), wishList.getMenu());
+    public WishListResponse mapWishListToWishListResponse(WishList wishList) {
+        return new WishListResponse(wishList.getId(), wishList.getMenu().getName(),wishList.getMenu().getPrice(),wishList.getMenu().getImageUrl());
     }
 }
